Removed V0.4. Reorganized files
